+++ src/hocs/index.js
@@ -1,0 +1,3 @@
+export { default as withLastSearchTerms } from './withLastSearchTerms/withLastSearchTerms';
+export { default as withLocation } from './withLocation/withLocation';
+export { withUseResourcesIds } from './withUseResourcesIds/withUseResourcesIds';