+++ src/hooks/index.js
@@ -25,3 +25,4 @@ export * from './useCallNumberTypesQuery';
 export * from './useAuditSettings';
 export * from './useCallNumberBrowseConfig';
 export * from './useSharedInstancesQuery';
+export * from './useResourcesIds';