+++ src/routes/buildManifestObject.js
@@ -20,34 +20,6 @@ export function buildManifestObject() {
     resultOffset: { initialValue: 0 },
     requestUrlQuery: { initialValue: '' },
     records: buildRecordsManifest(),
-    recordsToExportIDs: {
-      type: 'okapi',
-      records: 'ids',
-      accumulate: true,
-      fetch: false,
-      path: 'search/instances/ids',
-      throwErrors: false,
-      GET: {
-        params: {
-          query: buildSearchQuery(),
-        },
-        staticFallback: { params: {} },
-      },
-    },
-    holdingsToExportIDs: {
-      type: 'okapi',
-      records: 'ids',
-      accumulate: true,
-      fetch: false,
-      path: 'search/holdings/ids',
-      throwErrors: false,
-      GET: {
-        params: {
-          query: buildSearchQuery(),
-        },
-        staticFallback: { params: {} },
-      },
-    },
     quickExport: {
       type: 'okapi',
       fetch: false,